connect-ai/backend-python/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pymongo
python-dotenv
pandas
numpy
openai
httpx